server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# Vectorstore
pinecone==3.2.2
langchain-pinecone==0.1.2

# Embeddings
langchain-google-genai
google-genai

# PDF Parsing
pypdf

# Vision and Image Processing
pillow

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Logging
loguru

# --- Block 1: Database ---
sqlalchemy
psycopg2-binary  # The Python driver for PostgreSQL
alembic         # For database migrations (production best practice)

passlib[bcrypt]
python-jose[cryptography]

celery
redis

asyncio_throttle